Show images in a JLabel
Imported an image to a package and set it manually as an icon in a JLabel.
Resize the image to fit the width or height of the JLabel without modifying the aspect ratio.

Created a new project called Images, and a new JForm. On the designer menu, added 2 labels and deleted the text. Then imported 2 images from the folder to the project using the icon property of the labels. After that, added a buton for each image to resize them to the label's size. On this button we added a code to get both the label height and width and send it to a method that returned the imageIcon that we had to add to the setIcon method of the lbl. The method we used was called resizeImageIcon, in which we checked whether the height or the width of the original image was bigger. Declared an aspect ratio, took the desired height/width of the label depending on which was bigger and calculated the other one based on the aspect ratio. Ex: if width was bigger, desired height = width/ aspec ratio, else if height was bigger, desired width = height * aspec ratio.
Then the new and resized image was created using Image and its method getScaledInstance to scale the image. Then BufferedImage with its method getGraphics().drawImage to draw the image. And then convert it to imageicon to send it to the label and set it to the property icon.